+++ src/main/java/dad/us/dadVertx/entities/chat/ChatMessageState.java
@@ -12,13 +12,13 @@ public class ChatMessageState {
 
 	private Integer state_id;
 	private Integer message_id;
-	private Integer user_id;
+	private Long user_id;
 	private String state;
 	private Long timestamp;
 
 	@JsonCreator
 	public ChatMessageState(@JsonProperty("state_id") Integer state_id, @JsonProperty("message_id") Integer message_id,
-			@JsonProperty("user_id") Integer user_id, @JsonProperty("state") String state,
+			@JsonProperty("user_id") Long user_id, @JsonProperty("state") String state,
 			@JsonProperty("timestamp") Long timestamp) {
 		this.state_id = state_id;
 		this.message_id = message_id;
@@ -62,11 +62,11 @@ public class ChatMessageState {
 		this.message_id = message_id;
 	}
 
-	public Integer getUser_id() {
+	public Long getUser_id() {
 		return user_id;
 	}
 
-	public void setUser_id(Integer user_id) {
+	public void setUser_id(Long user_id) {
 		this.user_id = user_id;
 	}
