+++ src/main/java/dad/us/dadVertx/entities/psychology/TestResponse.java
@@ -8,14 +8,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class TestResponse {
 
-	private Integer user_id;
+	private Long user_id;
 	private Integer test_id;
 	private Integer question_number;
 	private String response;
 	private Long timestamp;
 
 	@JsonCreator
-	public TestResponse(@JsonProperty("user_id") Integer user_id, @JsonProperty("test_id") Integer test_id,
+	public TestResponse(@JsonProperty("user_id") Long user_id, @JsonProperty("test_id") Integer test_id,
 			@JsonProperty("question_number") Integer question_number, @JsonProperty("response") String response,
 			@JsonProperty("timestamp") Long timestamp) {
 		this.user_id = user_id;
@@ -25,11 +25,11 @@ public class TestResponse {
 		this.response = response;
 	}
 
-	public Integer getUser_id() {
+	public Long getUser_id() {
 		return user_id;
 	}
 
-	public void setUser_id(Integer user_id) {
+	public void setUser_id(Long user_id) {
 		this.user_id = user_id;
 	}
